requirements.txt
langchain==0.0.233
matplotlib==3.7.2
pandas==2.0.3
python-dotenv==1.0.0
streamlit==1.24.1
textract==1.6.5
transformers==4.30.2
pypdf
openai
tiktoken
faiss-cpu
duckduckgo-search